New final
Realized I was taking the incorrect approach to taking average values. I was attempted to see statistically how high n needed to be to guarantee the desired precision 100% of the time. Realized though that this required extensive computing power and time. Went back to recording number of points when the estimated value first equals the true pi value and then taking the average of that